refactor autograder components to reduce length
